Users/Gowrish/Desktop/InnoFusion 2/requirements.txt
Flask==2.3.3
pymongo==4.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-PyMongo==2.3.0
pytest==7.4.0
gunicorn==21.2.0
dnspython==2.4.2
gevent==22.10.2
psutil==5.9.5
pymongo[srv]==4.5.0
redis==4.5.5
certifi>=2023.7.22
pyopenssl>=23.2.0
cryptography>=41.0.3
